[commhistoryd] Avoid replaying notifications on startup. JB#62570
On startup Commhistoryd fetches and resolves existing notifications
from Commhistoryd. The deserialization of notification data calls
various setters and those marked the PersonalNotification instance
with hasPendingEvents, and then NotificationManager republished
the notifications because of that, effectively showing an notification
popup and playing feedback sound.

Made here the restored notification as not pending for any publication.
Also adjusted NotificationManager::slotContactResolveFinished()
not to have pending changes just because of related recipient being
resolved. It's not by itself yet a change to be published.